add auth form validation

+++ frontend/src/components/Auth.jsx
@@ -7,36 +7,51 @@ export function Auth({type, toggle}) {
     const [username, setUsername] = React.useState('');
     const [password, setPassword] = React.useState('');
     const [isLoading, setIsLoading] = React.useState(false);
+    const [isError, setIsError] = React.useState(false);
+    const [errorMessage, setErrorMessage] = React.useState("");
 
     const clear = () => {
         setUsername("");
         setPassword("");
-        window.location.reload(false);
     }
 
     const handleOnSubmit = async () => {
         setIsLoading(true);
         if (type === "sign up") {
             const response = await sign_up(username, password);
-            type = "sign in";
+            console.log(response);
+            if('name' in response) {
+              setIsLoading(false);
+              setIsError(true);
+              setErrorMessage(response.message ? response.message : "Username cannot be longer than 150 characters");
+            } else {
+              type = "sign in";
+            }
         }
         if (type === "sign in") {
             const response = await sign_in(username, password);
             setIsLoading(false);
-            window.localStorage.clear();
-            window.localStorage.setItem('access_token', response.access);
-            window.localStorage.setItem('refresh_token', response.refresh);
-            window.localStorage.setItem("isLogged", true);
-            window.localStorage.setItem("view", "Dashboard");
-            window.localStorage.setItem("username", jwtDecode(response.access).username);
-            clear();
+            if('name' in response) {
+              setIsLoading(false); 
+              setIsError(true); 
+              setErrorMessage(response.message ? response.message : "Username cannot be longer than 150 characters");
+            } else {
+              window.localStorage.clear();
+              window.localStorage.setItem('access_token', response.access);
+              window.localStorage.setItem('refresh_token', response.refresh);
+              window.localStorage.setItem("isLogged", true);
+              window.localStorage.setItem("view", "Dashboard");
+              window.localStorage.setItem("username", jwtDecode(response.access).username);
+              clear();
+              window.location.reload(false);
+            }   
         }
       };
 
     const handleToggle = () => {
-        clear();
-        if(type === "sign up") { toggle("sign in") }
-        else { toggle("sign up") }
+      clear();
+      if(type === "sign up") { setIsError(false); toggle("sign in") }
+      else { setIsError(false); toggle("sign up") }
     }
 
   return (
@@ -45,13 +60,15 @@ export function Auth({type, toggle}) {
         <div className="mb-2 block">
           <Label htmlFor="name" value="Your name" className="text-4xl" />
         </div>
-        <TextInput className="text-4xl" size="lg" id="name" type="name" placeholder="username" required shadow value={username} onChange={e => setUsername(e.target.value)} />
+        <TextInput className="text-4xl" size="lg" id="name" type="name" placeholder="username" required shadow value={username} onChange={e => {setIsError(false); setUsername(e.target.value)}} />
+        {isError && <Label value={errorMessage} className="text-sm" color="failure" />}
       </div>
       <div>
         <div className="mb-2 block">
           <Label htmlFor="password" value="Your password" className="text-4xl"/>
         </div>
-        <TextInput className="text-4xl" id="password" type="password" placeholder="********" required shadow value={password} onChange={e => setPassword(e.target.value)} />
+        <TextInput className="text-4xl" id="password" type="password" placeholder="********" required shadow value={password} onChange={e => {setIsError(false); setPassword(e.target.value)}} />
+        {isError && <Label value={errorMessage} className="text-sm" color="failure" />}
       </div>
       <div className="flex items-center gap-2">
         <Label htmlFor="agree" className="flex">
